refactor(kernel): phase 2 — extract free-fn / cron-bridge / probe helpers
Move the cohesive bottom-of-file helpers out of kernel/mod.rs into
thematic sub-modules of `kernel`:

  mcp_summary.rs       – mcp_summary_cache_key, render_mcp_summary
  reviewer_sanitize.rs – sanitize_reviewer_line, sanitize_reviewer_block
  cron_script.rs       – cron_script_wake_gate, atomic_write_toml,
                         parse_wake_gate
  cron_bridge.rs       – `impl CronChannelSender for KernelCronBridge`,
                         CRON_EMPTY_OUTPUT_HEARTBEAT, cron_fan_out_targets,
                         cron_deliver_response
  provider_probe.rs    – probe_and_update_local_provider,
                         probe_all_local_providers_once

Mechanical, behavior-preserving move. mod.rs drops by ~636 lines (now
~18.2k). The struct `KernelCronBridge` itself stays in mod.rs; only its
trait impl moves. Free fns formerly visible only to mod.rs are now
`pub(super)` so the (still-resident) `LibreFangKernel` methods that call
them keep compiling. mod.rs adds a `use cron_bridge::{cron_deliver_response,
cron_fan_out_targets};` re-export so existing call sites in mod.rs and
`kernel::tests` resolve byte-for-byte.

Also restores `kernel_handle::self` alongside the prelude wildcard
import — phase 1 dropped the self alias as "newly unused", missing the
fact that `kernel::tests` uses it via `kernel_handle::ApprovalGate::...`.
Without this, `cargo test -p librefang-kernel --lib --no-run` fails.
