+++ crates/librefang-kernel/src/kernel/mod.rs
@@ -23,7 +23,10 @@ use librefang_runtime::agent_loop::{
 };
 use librefang_runtime::audit::AuditLog;
 use librefang_runtime::drivers;
-use librefang_runtime::kernel_handle::prelude::*;
+// `kernel_handle::self` is needed by `kernel::tests` (call sites like
+// `kernel_handle::ApprovalGate::resolve_user_tool_decision(...)`) —
+// keep the self alias alongside the prelude wildcard so tests.rs resolves.
+use librefang_runtime::kernel_handle::{self, prelude::*};
 use librefang_runtime::llm_driver::{
     CompletionRequest, CompletionResponse, DriverConfig, LlmDriver, LlmError, StreamEvent,
 };
@@ -54,6 +57,23 @@ use tracing::{debug, error, info, instrument, warn};
 /// retain access to `LibreFangKernel`'s private fields and inherent
 /// methods without any visibility surgery.
 mod handles;
+
+// Cohesive free-fn / non-`LibreFangKernel`-impl chunks pulled out of
+// this file in Phase 2 of the kernel/mod.rs split. Re-exported below so
+// existing call sites — including `super::foo` references from
+// `kernel::tests` — continue to resolve unchanged.
+mod cron_bridge;
+mod cron_script;
+mod mcp_summary;
+mod provider_probe;
+mod reviewer_sanitize;
+
+use cron_bridge::{cron_deliver_response, cron_fan_out_targets};
+use cron_script::{atomic_write_toml, cron_script_wake_gate};
+use mcp_summary::{mcp_summary_cache_key, render_mcp_summary};
+use provider_probe::probe_all_local_providers_once;
+pub use provider_probe::probe_and_update_local_provider;
+use reviewer_sanitize::{sanitize_reviewer_block, sanitize_reviewer_line};
 
 /// Synthetic `SenderContext.channel` value the cron dispatcher uses for
 /// `[[cron_jobs]]` fires. Matched in [`KernelHandle::resolve_user_tool_decision`]
@@ -17294,364 +17314,11 @@ enum ReviewError {
     Permanent(String),
 }
 
-/// Build a deterministic cache key for the per-agent MCP allowlist; sorts and joins with `\x1f` so insertion-order variants share one entry.
-fn mcp_summary_cache_key(mcp_allowlist: &[String]) -> String {
-    if mcp_allowlist.is_empty() {
-        return String::from("*");
-    }
-    let mut sorted = mcp_allowlist.to_vec();
-    sorted.sort();
-    sorted.join("\x1f")
-}
+// `mcp_summary_cache_key` and `render_mcp_summary` live in `kernel::mcp_summary`.
 
-/// Render the MCP-server tool summary that lands in the system prompt.
-///
-/// Pulled out of [`Kernel::build_mcp_summary`] so it can be unit-tested
-/// without instantiating a full kernel. Determinism is load-bearing:
-///
-/// - Servers are grouped in a `BTreeMap` so the outer iteration order is
-///   lexicographic, not HashMap-random across processes.
-/// - Each server's tool list is sorted before joining — `tools_in` carries
-///   MCP-server-connect order which varies run-to-run and would otherwise
-///   defeat provider prompt caching even when the underlying tool set is
-///   identical.
-///
-/// See issue #3298 and the regression test
-/// `tests::mcp_summary_is_byte_identical_across_input_orders` below.
-fn render_mcp_summary(
-    tools_in: &[String],
-    configured_servers: &[String],
-    mcp_allowlist: &[String],
-) -> String {
-    if tools_in.is_empty() {
-        return String::new();
-    }
+// `sanitize_reviewer_line` and `sanitize_reviewer_block` live in `kernel::reviewer_sanitize`.
 
-    let normalized: Vec<String> = mcp_allowlist
-        .iter()
-        .map(|s| librefang_runtime::mcp::normalize_name(s))
-        .collect();
-
-    let mut servers: std::collections::BTreeMap<String, Vec<String>> =
-        std::collections::BTreeMap::new();
-    let mut tool_count = 0usize;
-    for tool_name in tools_in {
-        if let Some(server_name) = librefang_runtime::mcp::resolve_mcp_server_from_known(
-            tool_name,
-            configured_servers.iter().map(String::as_str),
-        ) {
-            let normalized_server = librefang_runtime::mcp::normalize_name(server_name);
-            if !mcp_allowlist.is_empty() && !normalized.iter().any(|n| n == &normalized_server) {
-                continue;
-            }
-            if let Some(raw_tool_name) =
-                tool_name.strip_prefix(&format!("mcp_{}_", normalized_server))
-            {
-                servers
-                    .entry(normalized_server)
-                    .or_default()
-                    .push(raw_tool_name.to_string());
-            } else {
-                servers
-                    .entry(normalized_server)
-                    .or_default()
-                    .push(tool_name.clone());
-            }
-        } else {
-            servers
-                .entry("unknown".to_string())
-                .or_default()
-                .push(tool_name.clone());
-        }
-        tool_count += 1;
-    }
-    if tool_count == 0 {
-        return String::new();
-    }
-    // Sort each server's tool list so the rendered summary is byte-stable
-    // across processes — see function-level docs.
-    for tool_names in servers.values_mut() {
-        tool_names.sort();
-    }
-    let mut summary = format!("\n\n--- Connected MCP Servers ({} tools) ---\n", tool_count);
-    for (server, tool_names) in &servers {
-        summary.push_str(&format!(
-            "- {server}: {} tools ({})\n",
-            tool_names.len(),
-            tool_names.join(", ")
-        ));
-    }
-    summary.push_str("MCP tools are prefixed with mcp_{server}_ and work like regular tools.\n");
-    let has_filesystem = servers.keys().any(|s| s.contains("filesystem"));
-    if has_filesystem {
-        summary.push_str(
-            "IMPORTANT: For accessing files OUTSIDE your workspace directory, you MUST use \
-             the MCP filesystem tools (e.g. mcp_filesystem_read_file, mcp_filesystem_list_directory) \
-             instead of the built-in file_read/file_list/file_write tools, which are restricted to \
-             the workspace. The MCP filesystem server has been granted access to specific directories \
-             by the user.",
-        );
-    }
-    summary
-}
-
-/// Sanitize a single-line author-supplied string (skill name, description)
-/// for safe interpolation into the reviewer's user message.
-///
-/// Thin wrapper over `librefang_runtime::prompt_builder::sanitize_for_prompt`
-/// — delegating keeps the bracket- and control-char rules consistent with
-/// the main prompt builder.
-fn sanitize_reviewer_line(s: &str, max_chars: usize) -> String {
-    librefang_runtime::prompt_builder::sanitize_for_prompt(s, max_chars)
-}
-
-/// Sanitize a multi-line block (trace summary, response summary) for
-/// embedding inside `<data>…</data>` markers in the reviewer prompt.
-///
-/// Preserves `\n` (the caller wants readable structure) but strips:
-/// - `\r`, null bytes, and other C0 control characters that some LLMs
-///   misinterpret as structural separators.
-/// - Triple backticks, so the reviewer can't be tricked into treating
-///   content as the start of its own code-fenced answer block (which
-///   `extract_json_from_llm_response` later greps for).
-/// - `<data>` / `</data>` markers, so nothing inside the block can
-///   prematurely close our envelope and escape into instructional scope.
-///
-/// Hard-capped at `max_chars`; truncation is signalled with a trailing
-/// `" …[truncated]"`.
-fn sanitize_reviewer_block(s: &str, max_chars: usize) -> String {
-    let mut out = String::with_capacity(s.len().min(max_chars));
-    for ch in s.chars() {
-        // Keep \n, \t. Drop other controls. Everything else passes.
-        if ch == '\n' || ch == '\t' {
-            out.push(ch);
-        } else if ch.is_control() {
-            out.push(' ');
-        } else {
-            out.push(ch);
-        }
-    }
-    // Neutralize markers that could break out of the reviewer's data block
-    // or forge an answer code fence. Replace rather than strip so the
-    // content's shape (indentation, line structure) stays recognizable.
-    let out = out
-        .replace("```", "``")
-        .replace("<data>", "(data)")
-        .replace("</data>", "(/data)");
-    if out.chars().count() <= max_chars {
-        return out;
-    }
-    // UTF-8-safe truncation: keep chars, not bytes.
-    let truncated: String = out.chars().take(max_chars.saturating_sub(14)).collect();
-    format!("{truncated} …[truncated]")
-}
-
-/// Run a cron job's pre-check script and parse the wake gate from its output.
-///
-/// Returns `true` if the agent should be woken (normal path), `false` to skip.
-///
-/// Rules (mirrors Hermes `_parse_wake_gate`):
-/// - Script must exit 0; on any error we default to waking the agent.
-/// - Find the last non-empty stdout line and try to parse it as JSON.
-/// - If the parsed object has `"wakeAgent": false` (strict bool), return false.
-/// - Everything else (non-JSON, missing key, null, 0, "") → return true.
-///
-/// # Security hardening
-///
-/// `pre_check_script` used to inherit the full daemon environment, allowing
-/// it to read API keys and other secrets from env vars.  It also had no
-/// working-directory restriction and no stdout size limit.
-///
-/// This implementation now:
-/// * Clears the inherited environment with `env_clear()` so daemon secrets
-///   are not leaked to the child process.
-/// * Passes only `PATH` and `HOME` so the script can still locate standard
-///   binaries without receiving application-layer credentials.
-/// * Sets `current_dir` to the agent workspace when one is available,
-///   otherwise falls back to a system temp directory.
-/// * Caps stdout (and stderr) at 64 KiB to prevent a misbehaving script
-///   from filling daemon memory.
-async fn cron_script_wake_gate(
-    job_name: &str,
-    script_path: &str,
-    agent_workspace: Option<&std::path::Path>,
-) -> bool {
-    use std::process::Stdio;
-    use tokio::io::AsyncReadExt;
-    use tokio::process::Command;
-
-    /// Maximum bytes we read from stdout before truncating.
-    const MAX_OUTPUT: usize = 64 * 1024; // 64 KiB
-
-    // Resolve a safe working directory for the child process.
-    // Preference order: agent workspace → system temp → current dir.
-    let cwd = agent_workspace
-        .map(|p| p.to_path_buf())
-        .unwrap_or_else(std::env::temp_dir);
-
-    // Build the command with a stripped-down environment.
-    // `env_clear` prevents all inherited daemon env vars (API keys, secrets,
-    // socket paths, etc.) from reaching the child.  We selectively restore
-    // the two vars that most scripts need to function correctly.
-    let mut cmd = Command::new(script_path);
-    cmd.env_clear();
-    if let Ok(path_val) = std::env::var("PATH") {
-        cmd.env("PATH", path_val);
-    }
-    if let Ok(home_val) = std::env::var("HOME") {
-        cmd.env("HOME", home_val);
-    }
-    cmd.current_dir(&cwd);
-    cmd.stdout(Stdio::piped());
-    cmd.stderr(Stdio::piped());
-    cmd.kill_on_drop(true);
-
-    // Hard cap: pre-check scripts must complete within 30 s.
-    // A hung script would otherwise block the cron dispatcher indefinitely.
-    let run = async {
-        let child = cmd.spawn();
-        match child {
-            Err(e) => Err(e),
-            Ok(mut child) => {
-                // Cap stdout at MAX_OUTPUT bytes.
-                let mut stdout_buf = Vec::with_capacity(MAX_OUTPUT.min(4096));
-                if let Some(stdout) = child.stdout.take() {
-                    let _ = stdout
-                        .take(MAX_OUTPUT as u64)
-                        .read_to_end(&mut stdout_buf)
-                        .await;
-                }
-                // Drain stderr (up to the same cap) to avoid blocking the child.
-                if let Some(stderr) = child.stderr.take() {
-                    let mut _discard = Vec::new();
-                    let _ = stderr
-                        .take(MAX_OUTPUT as u64)
-                        .read_to_end(&mut _discard)
-                        .await;
-                }
-                let status = child.wait().await?;
-                Ok((status, stdout_buf))
-            }
-        }
-    };
-
-    let (status, raw_stdout) =
-        match tokio::time::timeout(std::time::Duration::from_secs(30), run).await {
-            Err(_elapsed) => {
-                tracing::warn!(
-                    job = %job_name,
-                    script = %script_path,
-                    "cron: pre-check script timed out after 30s, waking agent"
-                );
-                return true;
-            }
-            Ok(Err(e)) => {
-                tracing::warn!(
-                    job = %job_name,
-                    script = %script_path,
-                    error = %e,
-                    "cron: pre-check script failed to launch, waking agent"
-                );
-                return true;
-            }
-            Ok(Ok(pair)) => pair,
-        };
-
-    if !status.success() {
-        tracing::warn!(
-            job = %job_name,
-            script = %script_path,
-            code = ?status.code(),
-            "cron: pre-check script exited non-zero, waking agent"
-        );
-        return true;
-    }
-
-    let stdout = String::from_utf8_lossy(&raw_stdout);
-    parse_wake_gate(&stdout)
-}
-
-/// Atomically write a TOML file by staging the new content in a sibling
-/// `.tmp` file and renaming it over the destination.
-///
-/// SECURITY / CORRECTNESS: a plain `fs::write` is non-atomic. Two
-/// concurrent persisters (e.g. `patch_agent` + `set_agent_model`) can
-/// truncate each other's output mid-flight, and a process crash at the
-/// wrong moment leaves a half-written file that fails to parse on next
-/// boot. `rename` is atomic on POSIX filesystems and effectively atomic
-/// on Windows for files on the same volume; if the rename fails we
-/// clean up the staging file.
-///
-/// We also `sync_all` the temp file before rename so the bytes hit the
-/// disk before the directory entry is swapped — without that, a power
-/// loss could leave the renamed file pointing at empty/stale data even
-/// though the rename succeeded.
-fn atomic_write_toml(path: &Path, content: &str) -> std::io::Result<()> {
-    use std::io::Write;
-    use std::sync::atomic::{AtomicU64, Ordering};
-    // Per-call counter so two threads in the same process never share
-    // a tmp filename — otherwise concurrent writers can clobber each
-    // other's staging file before rename, defeating the atomicity.
-    static SEQ: AtomicU64 = AtomicU64::new(0);
-    let seq = SEQ.fetch_add(1, Ordering::Relaxed);
-
-    // Same-directory tmp path keeps rename on the same filesystem so
-    // it's a true atomic in-place swap rather than a cross-volume copy.
-    let mut tmp = path.to_path_buf();
-    let file_name = path
-        .file_name()
-        .ok_or_else(|| std::io::Error::new(std::io::ErrorKind::InvalidInput, "missing filename"))?
-        .to_os_string();
-    let mut tmp_name = file_name;
-    tmp_name.push(format!(".{}.{seq}.tmp", std::process::id()));
-    tmp.set_file_name(tmp_name);
-
-    let write_result = (|| -> std::io::Result<()> {
-        let mut f = std::fs::File::create(&tmp)?;
-        f.write_all(content.as_bytes())?;
-        // fsync so the bytes hit disk before we publish via rename;
-        // without this a power loss between rename and flush would
-        // leave the renamed file pointing at empty/garbage data.
-        f.sync_all()?;
-        Ok(())
-    })();
-
-    if let Err(e) = write_result {
-        let _ = std::fs::remove_file(&tmp);
-        return Err(e);
-    }
-
-    // POSIX `rename` is atomic. Windows `MoveFileEx` with
-    // REPLACE_EXISTING (which Rust's std uses) is effectively atomic
-    // for files on the same volume, though there is a brief window
-    // where readers may see ERROR_SHARING_VIOLATION on contention.
-    if let Err(e) = std::fs::rename(&tmp, path) {
-        let _ = std::fs::remove_file(&tmp);
-        return Err(e);
-    }
-    Ok(())
-}
-
-/// Parse the wake gate from script stdout.
-///
-/// Finds the last non-empty line, tries JSON-decode, checks `wakeAgent`.
-/// Returns `true` (wake) unless `wakeAgent` is strictly `false`.
-fn parse_wake_gate(script_output: &str) -> bool {
-    let last_line = script_output.lines().rfind(|l| !l.trim().is_empty());
-
-    let last_line = match last_line {
-        Some(l) => l.trim(),
-        None => return true,
-    };
-
-    let value: serde_json::Value = match serde_json::from_str(last_line) {
-        Ok(v) => v,
-        Err(_) => return true,
-    };
-
-    // Only `{"wakeAgent": false}` (strict bool false) skips the agent.
-    value.get("wakeAgent") != Some(&serde_json::Value::Bool(false))
-}
+// `cron_script_wake_gate`, `atomic_write_toml`, and the private `parse_wake_gate` helper live in `kernel::cron_script`.
 
 /// Adapter from the kernel's `send_channel_message` to the
 /// `CronChannelSender` trait used by the multi-target fan-out engine.
@@ -17659,168 +17326,7 @@ struct KernelCronBridge {
     kernel: Arc<LibreFangKernel>,
 }
 
-#[async_trait::async_trait]
-impl crate::cron_delivery::CronChannelSender for KernelCronBridge {
-    async fn send_channel_message(
-        &self,
-        channel_type: &str,
-        recipient: &str,
-        message: &str,
-        thread_id: Option<&str>,
-        account_id: Option<&str>,
-    ) -> Result<(), String> {
-        self.kernel
-            .send_channel_message(channel_type, recipient, message, thread_id, account_id)
-            .await
-            .map(|_| ())
-            .map_err(|e| e.to_string())
-    }
-}
-
-/// Sentinel body sent when the agent / workflow produced no output but the
-/// caller still wants every fan-out target invoked (heartbeat semantics).
-/// Plain text so all adapters render it identically.
-const CRON_EMPTY_OUTPUT_HEARTBEAT: &str = "(cron heartbeat: empty output)";
-
-/// Fan out `output` to every target in `delivery_targets` concurrently.
-///
-/// Best-effort: never returns an error, because the cron job itself has
-/// already succeeded by the time we get here. Per-target failures are
-/// counted and logged. The legacy single-destination `delivery` field is
-/// handled separately by [`cron_deliver_response`].
-///
-/// **Empty output is not silently dropped.** When `output.is_empty()` we
-/// substitute a short heartbeat marker so every configured target still
-/// fires — the previous early-return swallowed the delivery entirely and
-/// broke liveness-style cron jobs (e.g. "ping #ops every hour even when I
-/// have nothing to say"). Cron jobs that genuinely want to skip empty-
-/// output runs should not configure fan-out targets at all.
-async fn cron_fan_out_targets(
-    kernel: &Arc<LibreFangKernel>,
-    job_name: &str,
-    output: &str,
-    targets: &[librefang_types::scheduler::CronDeliveryTarget],
-) {
-    if targets.is_empty() {
-        return;
-    }
-    let payload: &str = if output.is_empty() {
-        CRON_EMPTY_OUTPUT_HEARTBEAT
-    } else {
-        output
-    };
-    let sender: Arc<dyn crate::cron_delivery::CronChannelSender> = Arc::new(KernelCronBridge {
-        kernel: kernel.clone(),
-    });
-    let engine = crate::cron_delivery::CronDeliveryEngine::new(sender);
-    let results = engine.deliver(targets, job_name, payload).await;
-    let total = results.len();
-    let failures = results.iter().filter(|r| !r.success).count();
-    let successes = total - failures;
-    if failures == 0 {
-        tracing::info!(
-            job = %job_name,
-            targets = total,
-            "Cron fan-out: all {successes} target(s) delivered"
-        );
-    } else {
-        tracing::warn!(
-            job = %job_name,
-            total = total,
-            ok = successes,
-            failed = failures,
-            "Cron fan-out: partial delivery"
-        );
-        for r in results.iter().filter(|r| !r.success) {
-            tracing::warn!(
-                job = %job_name,
-                target = %r.target,
-                error = %r.error.as_deref().unwrap_or(""),
-                "Cron fan-out: target failed"
-            );
-        }
-    }
-}
-
-/// Deliver a cron job's agent response to the configured delivery target.
-async fn cron_deliver_response(
-    kernel: &LibreFangKernel,
-    agent_id: AgentId,
-    response: &str,
-    delivery: &librefang_types::scheduler::CronDelivery,
-) {
-    use librefang_types::scheduler::CronDelivery;
-
-    if response.is_empty() {
-        return;
-    }
-
-    match delivery {
-        CronDelivery::None => {}
-        CronDelivery::Channel { channel, to } => {
-            tracing::debug!(channel = %channel, to = %to, "Cron: delivering to channel");
-            // Persist as last channel for this agent (survives restarts)
-            let kv_val = serde_json::json!({"channel": channel, "recipient": to});
-            let _ = kernel
-                .memory
-                .structured_set(agent_id, "delivery.last_channel", kv_val);
-            if let Err(e) = kernel
-                .send_channel_message(channel, to, response, None, None)
-                .await
-            {
-                tracing::warn!(channel = %channel, to = %to, error = %e, "Cron channel delivery failed");
-            }
-        }
-        CronDelivery::LastChannel => {
-            match kernel
-                .memory
-                .structured_get(agent_id, "delivery.last_channel")
-            {
-                Ok(Some(val)) => {
-                    let channel = val["channel"].as_str().unwrap_or("");
-                    let recipient = val["recipient"].as_str().unwrap_or("");
-                    if !channel.is_empty() && !recipient.is_empty() {
-                        tracing::info!(
-                            channel = %channel,
-                            recipient = %recipient,
-                            "Cron: delivering to last channel"
-                        );
-                        if let Err(e) = kernel
-                            .send_channel_message(channel, recipient, response, None, None)
-                            .await
-                        {
-                            tracing::warn!(channel = %channel, recipient = %recipient, error = %e, "Cron last_channel delivery failed");
-                        }
-                    }
-                }
-                _ => {
-                    tracing::debug!("Cron: no last channel found for agent {}", agent_id);
-                }
-            }
-        }
-        CronDelivery::Webhook { url } => {
-            tracing::debug!(url = %url, "Cron: delivering via webhook");
-            let client = librefang_runtime::http_client::proxied_client_builder()
-                .timeout(std::time::Duration::from_secs(30))
-                .build();
-            if let Ok(client) = client {
-                let payload = serde_json::json!({
-                    "agent_id": agent_id.to_string(),
-                    "response": response,
-                    "timestamp": chrono::Utc::now().to_rfc3339(),
-                });
-                match client.post(url).json(&payload).send().await {
-                    Ok(resp) => {
-                        tracing::debug!(status = %resp.status(), "Cron webhook delivered");
-                    }
-                    Err(e) => {
-                        tracing::warn!(error = %e, "Cron webhook delivery failed");
-                    }
-                }
-            }
-        }
-    }
-}
+// `CronChannelSender` impl, `cron_fan_out_targets`, and `cron_deliver_response` live in `kernel::cron_bridge`. The `KernelCronBridge` struct definition stays here because it holds an `Arc<LibreFangKernel>` shared with the rest of the cron dispatcher.
 
 impl LibreFangKernel {
     /// Mark all active Hands' cron jobs as due-now so the next scheduler tick fires them.
@@ -18597,14 +18103,7 @@ impl LibreFangKernel {
 // for the `auth_status` of local providers (Ollama / vLLM / LM Studio /
 // lemonade) — no other code writes `NotRequired` or `LocalOffline` to them.
 
-/// Probe a single local provider and update its catalog entry.
-///
-/// Returns the probe result so callers (e.g. the test endpoint) can surface
-/// latency / error detail in their response.
-///
-/// `log_offline_as_warn = true` for providers in the default-or-fallback set
-/// (a real misconfiguration), `false` for incidentally-defined local
-/// providers (not configured — expected to be offline).
+// `probe_and_update_local_provider` and `probe_all_local_providers_once` live in `kernel::provider_probe`. The inherent `LibreFangKernel::probe_local_provider` method-style facade stays here.
 impl LibreFangKernel {
     /// Method-style facade over [`probe_and_update_local_provider`] so callers
     /// outside this crate (e.g. `librefang-api`) do not need to import the
@@ -18618,138 +18117,6 @@ impl LibreFangKernel {
     ) -> librefang_runtime::provider_health::ProbeResult {
         probe_and_update_local_provider(self, provider_id, base_url, log_offline_as_warn).await
     }
-}
-
-pub async fn probe_and_update_local_provider(
-    kernel: &Arc<LibreFangKernel>,
-    provider_id: &str,
-    base_url: &str,
-    log_offline_as_warn: bool,
-) -> librefang_runtime::provider_health::ProbeResult {
-    // Forward the provider's api_key (when configured) so reverse-proxy
-    // frontends like Open WebUI accept the listing request. Without this,
-    // the probe always 401s and the catalog flips to LocalOffline even
-    // when the underlying ollama is healthy.
-    let api_key = {
-        let catalog = kernel.model_catalog.load();
-        let env_var = catalog
-            .get_provider(provider_id)
-            .map(|p| p.api_key_env.clone())
-            .filter(|env| !env.trim().is_empty())
-            .unwrap_or_else(|| format!("{}_API_KEY", provider_id.to_uppercase().replace('-', "_")));
-        std::env::var(env_var).ok().filter(|v| !v.trim().is_empty())
-    };
-    let result = librefang_runtime::provider_health::probe_provider(
-        provider_id,
-        base_url,
-        api_key.as_deref(),
-    )
-    .await;
-    if result.reachable {
-        info!(
-            provider = %provider_id,
-            models = result.discovered_models.len(),
-            latency_ms = result.latency_ms,
-            "Local provider online"
-        );
-        // Pre-compute the merged info outside the RCU closure so it's not
-        // recomputed on retry (the closure may run multiple times if a
-        // concurrent updater wins the CAS).
-        let merged_info: Option<Vec<librefang_runtime::provider_health::DiscoveredModelInfo>> =
-            if result.discovered_models.is_empty() {
-                None
-            } else if result.discovered_model_info.is_empty() {
-                Some(
-                    result
-                        .discovered_models
-                        .iter()
-                        .map(
-                            |name| librefang_runtime::provider_health::DiscoveredModelInfo {
-                                name: name.clone(),
-                                parameter_size: None,
-                                quantization_level: None,
-                                family: None,
-                                families: None,
-                                size: None,
-                                capabilities: vec![],
-                            },
-                        )
-                        .collect(),
-                )
-            } else {
-                Some(result.discovered_model_info.clone())
-            };
-        kernel.model_catalog_update(|catalog| {
-            catalog.set_provider_auth_status(
-                provider_id,
-                librefang_types::model_catalog::AuthStatus::NotRequired,
-            );
-            if let Some(ref info) = merged_info {
-                catalog.merge_discovered_models(provider_id, info);
-            }
-        });
-    } else {
-        let err = result.error.as_deref().unwrap_or("unknown");
-        if log_offline_as_warn {
-            warn!(
-                provider = %provider_id,
-                error = err,
-                "Configured local provider offline"
-            );
-        } else {
-            debug!(
-                provider = %provider_id,
-                error = err,
-                "Local provider offline (not configured as default/fallback)"
-            );
-        }
-        // Mark unreachable local providers as LocalOffline (not Missing).
-        // Using Missing would cause detect_auth() to reset the status back
-        // to NotRequired on the next unrelated auth check, making offline
-        // providers reappear in the model switcher.
-        kernel.model_catalog_update(|catalog| {
-            catalog.set_provider_auth_status(
-                provider_id,
-                librefang_types::model_catalog::AuthStatus::LocalOffline,
-            );
-        });
-    }
-    result
-}
-
-/// Probe every local provider once and update the catalog. Called from the
-/// periodic loop in `start_background_agents`.
-///
-/// Probes run concurrently via `join_all`. The total wall time of one cycle
-/// is bounded by the slowest probe (≤ 2 s per provider — see
-/// `PROBE_TIMEOUT_SECS` in `provider_health`) instead of the sum across
-/// providers, which matters when a local server is hung rather than simply
-/// offline.
-async fn probe_all_local_providers_once(
-    kernel: &Arc<LibreFangKernel>,
-    relevant_providers: &std::collections::HashSet<String>,
-) {
-    let local_providers: Vec<(String, String)> = {
-        let catalog = kernel.model_catalog.load();
-        catalog
-            .list_providers()
-            .iter()
-            .filter(|p| {
-                librefang_runtime::provider_health::is_local_provider(&p.id)
-                    && !p.base_url.is_empty()
-            })
-            .map(|p| (p.id.clone(), p.base_url.clone()))
-            .collect()
-    };
-    let tasks = local_providers.into_iter().map(|(provider_id, base_url)| {
-        let kernel = Arc::clone(kernel);
-        let is_relevant = relevant_providers.contains(&provider_id.to_lowercase());
-        async move {
-            let _ = probe_and_update_local_provider(&kernel, &provider_id, &base_url, is_relevant)
-                .await;
-        }
-    });
-    futures::future::join_all(tasks).await;
 }
 
 // --- OFP Wire Protocol integration ---
